helloapp/requirements.txt
astroid==1.6.5
Django==2.1.1
django-filter==2.0.0
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pylint==1.9.3
pytz==2018.5
six==1.11.0
wrapt==1.10.11